rename : 테스트 용 코드 삭제 및 directory 위치 변경
